Stub in 'imageset' data source, keeping 'psd'
* Arguments differentiate between psd and imageset
* Add config file for imageset layers
* Add image files for each layer, reusing similar images
* Refactor PSD code to its own file
