+++ constants.py
@@ -2,3 +2,7 @@
 
 HTML_BREAK_PATTERN = r"<br/?>"
 HTML_BREAK = "<br>"
+
+# TODO save to user-editable file for customizability
+# Renders images sized to the bounding box of this layer
+BG_LAYER_NAME: str = "BG"
