+++ qunmk2_imageset.csv
@@ -1,0 +1,21 @@
+image_file,layer_name,x_pos,y_pos,width,height
+bg.jpg,BG,0,0,1856,1236
+s.png,s,1600,200
+o.png,1,50,500
+o.png,2,250,500
+o.png,3,450,500
+o.png,4,650,500
+o.png,5,850,500
+o.png,6,1050,500
+o.png,7,1250,500
+o.png,8,1450,500
+d.png,d,0,0
+o.png,n,450,200
+o.png,o,650,200
+o.png,lplay,300,800
+o.png,lr,300,800
+o.png,ls,300,800
+o.png,mplay,300,800
+o.png,param,300,800
+o.png,sys,300,800
+o.png,splay,0,800
